--- pkg/config/server.go
@@ -1,5 +1,0 @@
-package config
-
-type ServerConfiguration struct {
-	Port int `yaml:"port"`
-}
