+++ repository/db_init.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"fmt"
+
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 )
@@ -10,17 +11,17 @@ var db *gorm.DB
 
 func Init() error {
 	var err error
-	username := "root"           //账号
-	password := "" //密码
-	host := ""     //数据库地址，可以是Ip或者域名
-	port := 3306                 //数据库端口
-	Dbname := "douyin"           //数据库名
-	timeout := "10s"             //连接超时，10秒
+	username := "root" //账号
+	password := ""     //密码
+	host := ""         //数据库地址，可以是Ip或者域名
+	port := 3306       //数据库端口
+	Dbname := "douyin" //数据库名
+	timeout := "10s"   //连接超时，10秒
 	dsn := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?charset=utf8mb4&parseTime=True&loc=Local&timeout=%s", username, password, host, port, Dbname, timeout)
 	db, err = gorm.Open(mysql.Open(dsn), &gorm.Config{})
 	if err != nil {
 		panic(err)
 	}
-	err = db.AutoMigrate(&Following{}, &User{}, &Message{})
+	err = db.AutoMigrate(&Following{}, &User{}, &Message{}, &Comment{})
 	return err
 }
